refactor(accuracy): move BFCL gate-metric map to evaluation/bfcl_v4_metrics.py
Resolves the review thread on ACCURACY_METRIC_KEYS living in the generic
accuracy_results.py breakdown helper. The dict maps ruleset golden-metric names
(bfcl_*) to the BFCL scorer's breakdown keys — both halves are BFCL-specific, so
it doesn't belong in a scorer-agnostic module.

Move it to a new dependency-light evaluation/bfcl_v4_metrics.py (no bfcl-eval
import) so compliance/checker.py and metrics/results_plots.py import it without
pulling in the heavy BFCL scorer or the optional bfcl extra. Behavior-preserving;
the metric-keys test moves to test_bfcl_v4_metrics.py.
