+++ tests/unit/evaluation/test_bfcl_v4_metrics.py
@@ -1,0 +1,28 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for the BFCL v4 compliance-gate metric mapping."""
+
+import pytest
+from inference_endpoint.evaluation.bfcl_v4_metrics import ACCURACY_METRIC_KEYS
+
+
+@pytest.mark.unit
+def test_metric_keys():
+    assert ACCURACY_METRIC_KEYS["bfcl_overall_accuracy"] == "overall_accuracy"
+    assert (
+        ACCURACY_METRIC_KEYS["bfcl_normalized_accuracy"]
+        == "normalized_single_turn_score"
+    )
